+++ packages/key-facts/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.2.0](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.1.57...@times-components/key-facts@2.2.0) (2022-02-28)
+
+
+### Features
+
+* **TDP-1139:** live news analytics for in article components ([#2862](https://github.com/newsuk/times-components/issues/2862)) ([89f912b](https://github.com/newsuk/times-components/commit/89f912be1ac9e16a439d28ef93f905e59971303f))
+
+
+
+
+
 ## [2.1.57](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.1.56...@times-components/key-facts@2.1.57) (2022-02-16)
 
 **Note:** Version bump only for package @times-components/key-facts